+++ test/PerpetualIntelligence.Terminal.Test/Commands/Extractors/UnicodeChineseCommandExtractorTests.cs
@@ -144,7 +144,7 @@ namespace PerpetualIntelligence.Terminal.Commands.Extractors
 
         protected override void OnTestInitialize()
         {
-            options = MockTerminalOptions.NewOptions();
+            options = MockTerminalOptions.NewAliasOptions();
             textHandler = new UnicodeTextHandler();
             argExtractor = new OptionExtractor(textHandler, options, TestLogger.Create<OptionExtractor>());
             commands = new InMemoryCommandStore(textHandler, MockCommands.UnicodeCommands, options, TestLogger.Create<InMemoryCommandStore>());